+++ R/allScaffolds.R
@@ -1,0 +1,9 @@
+#' @title  All scaffolds
+#' @source /inst/script/allScaffolds.R
+#' @description This is a dataframe of all available scaffolds
+#' @docType data
+#' @usage data("allScaffolds", package = "spaceRATScaffolds")
+#' @format a data.frame
+#' @return a data.frame
+#' @keywords datasets
+"allScaffolds"